Fixed WEM duplication.

+++ AlienIsolationAudioParser/Program.cs
@@ -14,105 +14,92 @@ namespace AlienIsolationAudioExtractor
     class Program
     {
         private static string[] directories = { "DATA\\SOUND", "DATA\\SOUND_ORGANISED", "DATA\\SOUND_UNORGANISED" };
-        private enum CounterType { EXPORTED, NOT_FOUND, SKIPPED };
 
         static void Main(string[] args)
         {
-            try
+            Console.ForegroundColor = ConsoleColor.White;
+            Console.WriteLine("Alien: Isolation Audio Extractor" + Environment.NewLine +
+                                "Created by Matt Filer" + Environment.NewLine +
+                                "---" + Environment.NewLine +
+                                "This process will take around 30 minutes to complete." + Environment.NewLine +
+                                "---");
+
+            //Validate we're running in the correct directory
+            if (!File.Exists("AI.exe"))
             {
-                Console.ForegroundColor = ConsoleColor.White;
-                Console.WriteLine("Alien: Isolation Audio Extractor" + Environment.NewLine +
-                                  "Created by Matt Filer" + Environment.NewLine +
-                                  "---" + Environment.NewLine +
-                                  "This process will take around 30 minutes to complete." + Environment.NewLine +
-                                  "---");
-
-                //Validate we're running in the correct directory
-                if (!File.Exists("AI.exe"))
-                {
-                    Console.BackgroundColor = ConsoleColor.Yellow;
-                    Console.ForegroundColor = ConsoleColor.Black;
-                    Console.WriteLine("Could not locate sounds." + Environment.NewLine + "Please place this tool in your Alien: Isolation directory.");
-                    Console.Read();
-                    Environment.Exit(0);
-                }
-
-                //Check there are no lingering files which could interfere with export
-                if (Directory.Exists(directories[1]) || Directory.Exists(directories[2]))
-                {
-                    Console.BackgroundColor = ConsoleColor.Yellow;
-                    Console.ForegroundColor = ConsoleColor.Black;
-                    Console.WriteLine("Files remain from a previous conversion." + Environment.NewLine + "Please delete/rename folders \"" + directories[1] + "\" and \"" + directories[2] + "\" before running again.");
-                    Console.Read();
-                    //Environment.Exit(0);
-                }
-
-                //Wait for user input to start
-                Console.WriteLine(Environment.NewLine + "Press any key to begin.");
-                Console.Read();
-
-                //Copy resources and make directories
-                Directory.CreateDirectory(directories[1]);
-                Directory.CreateDirectory(directories[2]);
-                File.WriteAllBytes(directories[2] + "\\ww2ogg.exe", Properties.Resources.ww2ogg);
-                File.WriteAllBytes(directories[2] + "\\revorb.exe", Properties.Resources.revorb);
-                File.WriteAllBytes(directories[2] + "\\bnkextr.exe", Properties.Resources.bnkextr);
-                File.WriteAllBytes(directories[2] + "\\base_library.zip", Properties.Resources.base_library);
-                File.WriteAllBytes(directories[2] + "\\python36.dll", Properties.Resources.python36);
-                File.WriteAllBytes(directories[2] + "\\packed_codebooks_aoTuV_603.bin", Properties.Resources.packed_codebooks_aoTuV_603);
-                
-                //Extract and convert all files
-                Console.WriteLine("---" + Environment.NewLine + "Extracting soundbanks:");
-                extractSoundbanks();
-                Console.ForegroundColor = ConsoleColor.White;
-                Console.WriteLine(Environment.NewLine + "---" + Environment.NewLine + "Converting soundbanks:");
-                convertSoundbankFiles();
-                Console.ForegroundColor = ConsoleColor.White;
-                Console.WriteLine(Environment.NewLine + "---" + Environment.NewLine + "Converting remaining sound files:");
-                convertBaseSoundFiles();
-
-                //Clear up conversion resources
-                foreach (var file in Directory.GetFiles(directories[2], "*.exe", SearchOption.AllDirectories))
-                {
-                    File.Delete(file);
-                }
-                File.Delete(directories[2] + "\\base_library.zip");
-                File.Delete(directories[2] + "\\python36.dll");
-                File.Delete(directories[2] + "\\packed_codebooks_aoTuV_603.bin");
-
-                //Delete any empty directories from conversion
-                foreach (var directory in Directory.GetDirectories(directories[2]))
-                {
-                    if (Directory.GetFiles(directory).Length == 0 &&
-                        Directory.GetDirectories(directory).Length == 0)
-                    {
-                        Directory.Delete(directory, false);
-                    }
-                }
-
-                //Process file names
-                Console.ForegroundColor = ConsoleColor.White;
-                Console.WriteLine(Environment.NewLine + "---" + Environment.NewLine + "Starting file rename process.");
-                nameFilesIfPossible();
-
-                //Finished
-                Console.ForegroundColor = ConsoleColor.White;
-                Console.Write(Environment.NewLine + "---" + Environment.NewLine +
-                              "Finished processing all files." + Environment.NewLine +
-                              "Press any key to close.");
-                Console.ReadLine();
-                Environment.Exit(0);
-            }
-            catch
-            {
-                //Unknown unrecoverable error
-                Console.WriteLine("---");
                 Console.BackgroundColor = ConsoleColor.Yellow;
                 Console.ForegroundColor = ConsoleColor.Black;
-                Console.WriteLine("An unknown error occured during export.");
-                Console.ReadLine();
+                Console.WriteLine("Could not locate sounds." + Environment.NewLine + "Please place this tool in your Alien: Isolation directory.");
+                Console.Read();
                 Environment.Exit(0);
             }
+
+            //Check there are no lingering files which could interfere with export
+            if (Directory.Exists(directories[1]) || Directory.Exists(directories[2]))
+            {
+                Console.BackgroundColor = ConsoleColor.Yellow;
+                Console.ForegroundColor = ConsoleColor.Black;
+                Console.WriteLine("Files remain from a previous conversion." + Environment.NewLine + "Please delete/rename folders \"" + directories[1] + "\" and \"" + directories[2] + "\" before running again.");
+                Console.Read();
+                Environment.Exit(0);
+            }
+
+            //Wait for user input to start
+            Console.WriteLine(Environment.NewLine + "Press any key to begin.");
+            Console.Read();
+
+            //Copy resources and make directories
+            Directory.CreateDirectory(directories[1]);
+            Directory.CreateDirectory(directories[2]);
+            File.WriteAllBytes(directories[2] + "\\ww2ogg.exe", Properties.Resources.ww2ogg);
+            File.WriteAllBytes(directories[2] + "\\revorb.exe", Properties.Resources.revorb);
+            File.WriteAllBytes(directories[2] + "\\bnkextr.exe", Properties.Resources.bnkextr);
+            File.WriteAllBytes(directories[2] + "\\base_library.zip", Properties.Resources.base_library);
+            File.WriteAllBytes(directories[2] + "\\python36.dll", Properties.Resources.python36);
+            File.WriteAllBytes(directories[2] + "\\packed_codebooks_aoTuV_603.bin", Properties.Resources.packed_codebooks_aoTuV_603);
+                
+            //Extract and convert all files
+            Console.WriteLine("---" + Environment.NewLine + "Extracting soundbanks:");
+            extractSoundbanks();
+            Console.ForegroundColor = ConsoleColor.White;
+            Console.WriteLine(Environment.NewLine + "---" + Environment.NewLine + "Converting soundbanks:");
+            convertSoundbankFiles();
+            Console.ForegroundColor = ConsoleColor.White;
+            Console.WriteLine(Environment.NewLine + "---" + Environment.NewLine + "Converting remaining sound files:");
+            convertBaseSoundFiles();
+
+            //Clear up conversion resources
+            foreach (var file in Directory.GetFiles(directories[2], "*.exe", SearchOption.AllDirectories))
+            {
+                File.Delete(file);
+            }
+            File.Delete(directories[2] + "\\base_library.zip");
+            File.Delete(directories[2] + "\\python36.dll");
+            File.Delete(directories[2] + "\\packed_codebooks_aoTuV_603.bin");
+
+            //Delete any empty directories from conversion
+            foreach (var directory in Directory.GetDirectories(directories[2]))
+            {
+                if (Directory.GetFiles(directory).Length == 0 &&
+                    Directory.GetDirectories(directory).Length == 0)
+                {
+                    Directory.Delete(directory, false);
+                }
+            }
+
+            //Process file names
+            Console.ForegroundColor = ConsoleColor.White;
+            Console.WriteLine(Environment.NewLine + "---" + Environment.NewLine + "Renaming files where possible:");
+            nameFilesIfPossible();
+
+            //Finished
+            Console.ForegroundColor = ConsoleColor.White;
+            Console.Write(Environment.NewLine + "---" + Environment.NewLine +
+                            "Finished processing all files." + Environment.NewLine +
+                            "Press any key to close.");
+            Console.ReadLine();
+            Console.ReadLine();
+            Environment.Exit(0);
         }
 
 
@@ -124,11 +111,14 @@ namespace AlienIsolationAudioExtractor
             //Load soundbank & files listed
             XDocument soundbankXML = XDocument.Parse(Properties.Resources.SOUNDBANKSINFO);
             var fileArray = soundbankXML.Descendants("File");
-            int[] counter = { 0,0,0 }; 
 
             //Run through each file...
+            int counter = 0;
             foreach (var file in fileArray)
             {
+                counter++;
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.Write("\r{0}   ", counter + " of " + fileArray.Count() + " processed.");
                 try
                 {
                     //Get attributes of current file
@@ -146,50 +136,49 @@ namespace AlienIsolationAudioExtractor
                     //Skip out any non sfx files - currently unsupported
                     if (originalFileNameLanguage != "SFX")
                     {
-                        counter[(int)CounterType.SKIPPED]++;
                         continue;
                     }
 
                     //Try and find OGG file to rename
-                    bool hasFoundFile = false;
                     foreach (string oggFile in Directory.GetFiles(directories[2], originalFileNameID + ".ogg", SearchOption.AllDirectories))
                     {
                         if (!File.Exists(directories[1] + "\\SFX\\" + originalFileNameShortName))
                         {
                             File.Copy(oggFile, directories[1] + "\\SFX\\" + originalFileNameShortName);
                             File.Delete(oggFile);
-                            counter[(int)CounterType.EXPORTED]++;
+                            continue;
                         }
-                        hasFoundFile = true;
                     }
 
                     //Didn't find an OGG file, try find a WEM file (last resort)
-                    if (!hasFoundFile)
+                    foreach (string wemFile in Directory.GetFiles(directories[2], originalFileNameID + ".wem", SearchOption.AllDirectories))
                     {
-                        foreach (string wemFile in Directory.GetFiles(directories[2], originalFileNameID + ".wem", SearchOption.AllDirectories))
+                        if (!File.Exists(directories[1] + "\\SFX\\" + originalFileNameShortName.Substring(0, originalFileNameShortName.Length - 3) + "wem"))
                         {
-                            if (!File.Exists(directories[1] + "\\SFX\\" + originalFileNameShortName.Substring(0, originalFileNameShortName.Length - 3) + "wem"))
-                            {
-                                File.Copy(wemFile, directories[1] + "\\SFX\\" + originalFileNameShortName.Substring(0, originalFileNameShortName.Length - 3) + "wem");
-                                File.Delete(wemFile);
-                                counter[(int)CounterType.EXPORTED]++;
-                            }
-                            hasFoundFile = true;
+                            File.Copy(wemFile, directories[1] + "\\SFX\\" + originalFileNameShortName.Substring(0, originalFileNameShortName.Length - 3) + "wem");
+                            File.Delete(wemFile);
+                            continue;
                         }
-                    }
-
-                    //Didn't find a file at all!
-                    if (!hasFoundFile)
-                    {
-                        counter[(int)CounterType.NOT_FOUND]++;
                     }
                 }
                 catch
                 {
                     //Most likely no shortname, this is normal.
                 }
-                Console.ForegroundColor = ConsoleColor.Green;
-                Console.Write("\r{0}   ", counter[(int)CounterType.EXPORTED] + " renamed, " + counter[(int)CounterType.NOT_FOUND] + " referenced but not found, " + counter[(int)CounterType.SKIPPED] + " skipped out.");
+            }
+            foreach (var file in Directory.GetFiles(directories[1], "*.wem", SearchOption.AllDirectories))
+            {
+                if (File.Exists(file.Substring(0, file.Length - 3) + "ogg"))
+                {
+                    File.Delete(file); //Delete any duplicate WEM files that have come across
+                }
+            }
+            foreach (var file in Directory.GetFiles(directories[2], "*.wem", SearchOption.AllDirectories))
+            {
+                if (File.Exists(file.Substring(0, file.Length - 3) + "ogg"))
+                {
+                    File.Delete(file); //Delete all duplicate WEMs in unorganised directory
+                }
             }
         }
 
@@ -221,8 +210,8 @@ namespace AlienIsolationAudioExtractor
             }
             foreach (string file in Directory.GetFiles(directories[0], "*.WEM", SearchOption.AllDirectories))
             {
-                string fileThatShouldExist = directories[2] + "\\" + Path.GetFileNameWithoutExtension(file) + "."; //Support for directories here?
-                if (!File.Exists(fileThatShouldExist + "ogg"))
+                string fileThatShouldExist = directories[2] + "\\" + Path.GetFileNameWithoutExtension(file) + "."; 
+                if (!File.Exists(fileThatShouldExist + "ogg") && !File.Exists(fileThatShouldExist + "wem"))
                 {
                     File.Copy(file, fileThatShouldExist + "wem"); //Copy over any failed conversions to OGG in their raw WEM format to fill the gaps.
                 }
@@ -247,13 +236,6 @@ namespace AlienIsolationAudioExtractor
                 counter++;
                 Console.ForegroundColor = ConsoleColor.Green;
                 Console.Write("\r{0}   ", counter + " of " + searchQuery.Count() + " converted.");
-            }
-            foreach (var file in searchQuery)
-            {
-                if (File.Exists(file.Substring(0, file.Length - 3) + "ogg"))
-                {
-                    File.Delete(file); //Delete all converted WEMs (can't do this at conversion time as the file won't have processed yet)
-                }
             }
         }
